requirements.txt
smartapi-python==1.5.5
pyotp==2.9.0
python-dotenv==1.0.1
pandas==2.2.3
requests==2.32.3
APScheduler==3.10.4
tenacity==9.0.0
pandas_market_calendars==4.4.2
pycryptodome==3.21.0
websocket-client==1.8.0
logzero==1.7.0
